// src/app/shared/model/astro-month-report.model.ts
export class Astromonthreport{
    session_no: string;
    astro_name: string;
    logging_date: string;
    astroid: number;
    work_time: number;
    dnd_time: number;
    offline_time: number;
    online_time: number;
    total_call: number;
    received_call:number;
    missed_call:number;
    rejected_call: number;
    call_disconnect_normal_astro: number;
    call_disconnect_normal_user: number;
    call_disconnect_abrupt_astro: number;
    call_disconnect_abrupt_user:number;
    call_disconnect_astro_unreachable:number;
    call_disconnect_user_package_expired:number;
    call_disconnect_astro_not_pick: number;


}